requirements.txt
trio-websocket==0.11.*
pytest==7.4.*
pytest-trio==0.8.*